cleaned up all warnings

+++ src/ui.cpp
@@ -149,7 +149,7 @@ void DrawDebugInfo()
     if (!debugLines.empty())
     {
         float backgroundWidth = maxWidth + (padding * 2);
-        float backgroundHeight = (debugLines.size() * lineHeight) + padding;
+        float backgroundHeight = ((float)debugLines.size() * lineHeight) + padding;
         DrawRectangle(0, 0, (int)backgroundWidth, (int)backgroundHeight, {0, 0, 0, 120});
     }
 
@@ -159,7 +159,7 @@ void DrawDebugInfo()
         if (!debugLines[i].first.empty())
         {
             const char *text = debugLines[i].first.c_str();
-            DrawTextEx(fontSmallExtraLight, text, {padding, (i * lineHeight) + (padding / 2)}, (float)fontSmallExtraLight.baseSize, 2, WHITE);
+            DrawTextEx(fontSmallExtraLight, text, {padding, ((float)i * lineHeight) + (padding / 2)}, (float)fontSmallExtraLight.baseSize, 2, WHITE);
         }
     }
 
@@ -174,7 +174,7 @@ void DrawControlsInfo()
     const char *controls = "CTRL + R: Reload Game | F3: Debug | F10: Switch Monitor | F11: Borderless";
     Vector2 controlsSize = MeasureTextEx(fontSmallExtraLight, controls, (float)fontSmallExtraLight.baseSize, 2);
     DrawTextEx(fontSmallExtraLight, controls,
-               {screenWidth - controlsSize.x - 10, screenHeight - controlsSize.y - 10},
+               {(float)screenWidth - controlsSize.x - 10, (float)screenHeight - controlsSize.y - 10},
                (float)fontSmallExtraLight.baseSize, 2, WHITE);
 }
 
@@ -184,9 +184,9 @@ void DrawExitConfirmation()
     DrawRectangle(0, 0, screenWidth, screenHeight, {0, 0, 0, 200});
     const char *closeText = "close the window? (Y/n)";
     int fontSize = fontMedium.baseSize;
-    Vector2 textSize = MeasureTextEx(fontMedium, closeText, fontSize, 2.0f);
+    Vector2 textSize = MeasureTextEx(fontMedium, closeText, (float)fontSize, 2.0f);
     DrawTextEx(fontMedium, closeText,
-               {screenWidth / 2 - textSize.x / 2,
-                screenHeight / 2 - textSize.y / 2},
-               fontSize, 2.0f, WHITE);
+               {(float)screenWidth / 2 - textSize.x / 2,
+                (float)screenHeight / 2 - textSize.y / 2},
+               (float)fontSize, 2.0f, WHITE);
 }
